+++ src/models/filament.js
@@ -34,6 +34,10 @@ module.exports = function (app) {
             max: Number,
             experienced: Number
         },
+        printingSpeed: { // in mm/s
+            min: Number,
+            max: Number
+        },
         density: Number,
         buyDate: Date,
         price: Number,
@@ -43,7 +47,6 @@ module.exports = function (app) {
         initialTotalWeight: Number,  // in Kgram
         materialLeftPercentage: Number,
         flowPercentage: Number,
-        speedPercentage: Number,
         finished: { type: Boolean, default: false },
         finishedDate: { type: Date, default: null },
         _version : { type: Number }
@@ -61,8 +64,9 @@ module.exports = function (app) {
      * - 3: add lastUsedDate. Migrating to modificationDate.
      * - 4: add finished and finishedDate
      * - 5: add pricePerKG
+     * - 6: replace speedPercentage with printingSpeed range
      */
-    filamentSchema.statics.currentVersion = 5;
+    filamentSchema.statics.currentVersion = 6;
 
     filamentSchema.methods.getData = function(noPictures) {
         var data = this.toObject({getters: false, virtuals: true, versionKey: false});